feat: refactoring to path doubles
