+++ base/domain/path_management/path_doubles/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+from .base_path import BasePath
+
+
+__all__ = [
+    "BasePath",
+]
