+++ base/domain/path_management/path_doubles/base_path.py
