+++ base/domain/path_management/path_handler/__init__.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
 
-from .base_path import BasePath
 from .base_path_handler import BasePathHandler
 
 
 __all__ = [
-    "BasePath",
     "BasePathHandler",
 ]
