+++ base/infrastructure/path_management/path_doubles/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+from .path_faker import PathFaker
+
+
+__all__ = [
+    "PathFaker",
+]
