+++ base/infrastructure/path_management/path_handler/__init__.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
 
-from .fake_path import FakePath
 from .path_handler import PathHandler
 
 
 __all__ = [
-    "FakePath",
     "PathHandler",
 ]
