+++ tests/base/infrastructure/file_management/file_writer/test_yaml_file_writer.py
@@ -8,7 +8,7 @@ import pytest
 from base.infrastructure.file_management.file_handler import FakeFile
 from base.infrastructure.file_management.file_handler import FileHandler
 from base.infrastructure.file_management.file_writer import YamlFileWriter
-from base.infrastructure.path_management.path_handler import FakePath
+from base.infrastructure.path_management.path_doubles import PathFaker
 from base.infrastructure.path_management.path_handler import PathHandler
 
 # Domain
@@ -37,7 +37,7 @@ def test_yaml_file_writer_validation_with_valid_params():
 
     target_path = "/home/user1/project1/folder1"
     target_path = f"{target_path}/{fake_file.name}{fake_file.suffix}"
-    fake_path = FakePath(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
+    fake_path = PathFaker(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
     fake_path.touch()
 
     path_handler = PathHandler(path_obj=fake_path)
@@ -59,7 +59,7 @@ def test_yaml_file_writer_validation_with_invalid_path_not_exists():
 
     target_path = "/home/user1/project1/folder1"
     target_path = f"{target_path}/{fake_file.name}{fake_file.suffix}"
-    fake_path = FakePath(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
+    fake_path = PathFaker(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
 
     path_handler = PathHandler(path_obj=fake_path)
 
@@ -82,7 +82,7 @@ def test_yaml_file_writer_validation_with_invalid_file_type():
 
     target_path = "/home/user1/project1/folder1"
     target_path = f"{target_path}/{fake_file.name}{fake_file.suffix}"
-    fake_path = FakePath(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
+    fake_path = PathFaker(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
     fake_path.touch()
 
     path_handler = PathHandler(path_obj=fake_path)
